backend/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
python-dotenv==0.20.0
bcrypt==3.2.0
PyJWT==2.4.0
gunicorn==20.1.0
